Bug fixing and implementic logic changes
The bug was in the following. Method GetPageIdFromDataToken of GetPageId class was parsing page ID from entity token. But depending on different situations (add datafolder element, edit datafolder element, add metadata data) there will be expecting different entity tokens. I've implemented all these 3 cases, so now filtering for datafolders will work as well. If there are another cases except of datafolders and metadata fields, they should be also implemented.

Initially in the widget was an option to add page selector and to configure different filtering options: Level 1 (home page), level 2, level 3, level 4. But as we have to display a tree element to the end user, and a tree element should have some root, and on level 2 there could be 2,3,4,5 subtrees, there was no way to use these options level 2,3 and 4 and they didn't work as well. So there left only 1 available and possible for use option: level 1 - to show web pages only of some certain webpage. To leave a combobox with only 1 option "level 1" seemed not very good, therefore to PageReferenceSelectorWidgetFunctionBase instead of configuring option with levels was added a bool option: to filter or not to filter pages by active website. Because a standart bool widget provider was used, class HomePageSelectorWidgetFunction is not necessary anymore so it was deleted. As GetPageId returnt only ID of a root element, and all another code was removed, so I renamed it to GetHomePageId to avoid misunderstanding. GetRootPageId would be better as for me, but in all places was used a word "home", wherefore I used the same word and style